requirements.txt
APScheduler==3.8.1
autopep8==1.6.0
azure-common==1.1.27
azure-core==1.20.1
azure-nspkg==3.0.2
azure-storage-blob==12.9.0
azure-storage-file-share==12.6.0
backports.zoneinfo==0.2.1
boto3==1.21.13
botocore==1.24.13
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
Django==2.2.27
django-appconf==1.0.5
django-base64field==1.0
django-cors-headers==3.10.1
django-cryptography==1.0
django-rest-swagger==2.2.0
django-storages==1.11.1
djangorestframework==3.9.2
et-xmlfile==1.1.0
flake8==4.0.1
idna==3.3
install==1.3.4
isodate==0.6.0
itypes==1.2.0
Jinja2==3.1.1
jmespath==0.10.0
jsonfield==3.1.0
MarkupSafe==2.1.1
mccabe==0.6.1
msrest==0.6.21
mysqlclient==2.1.0
numpy==1.22.3
oauthlib==3.1.1
openapi-codec==1.3.2
openpyxl==3.0.9
pandas==1.3.4
protobuf==3.17.3
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
python-dateutil==2.8.2
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
requests-oauthlib==1.3.0
s3transfer==0.5.0
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.7
astroid==2.9.2
async-timeout==4.0.2
attrs==21.2.0
Automat==20.2.0
autopep8==1.6.0
beautifulsoup4==4.10.0
black==22.3.0
certifi==2020.11.8
chardet==3.0.4
click==8.1.2
constantly==15.1.0
Deprecated==1.2.13
Flask==1.1.2
Flask-Cors==3.0.10
gTTS==2.2.1
gunicorn==20.1.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
isort==5.10.1
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.7.1
lxml==4.6.3
MarkupSafe==1.1.1
mccabe==0.6.1
mutagen==1.45.1
mypy-extensions==0.4.3
numpy==1.21.2
packaging==21.3
pathspec==0.9.0
pep8==1.7.1
Pillow==8.0.1
platformdirs==2.4.1
pycodestyle==2.8.0
pylint==2.12.2
pyparsing==3.0.9
PyYAML==5.3.1
redis==4.3.3
requests==2.25.0
rq==1.10.1
selenium==3.141.0
six==1.15.0
soupsieve==2.2.1
supervisor==4.2.4
toml==0.10.2
tomli==2.0.1
Twisted==21.2.0
typing-extensions==4.0.1
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.13.3
zope.interface==5.4.0

